requirements.txt
fastapi
uvicorn
python-multipart
pandas
groq
python-dotenv
httpx
pydantic
google-genai
google-adk
